+++ backend/src/main/java/com/yigongil/backend/domain/study/studyquery/StudyQueryRepository.java
@@ -1,5 +1,6 @@
-package com.yigongil.backend.domain.study;
+package com.yigongil.backend.domain.study.studyquery;
 
+import com.yigongil.backend.domain.study.Study;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
