+++ backend/src/test/java/com/yigongil/backend/domain/study/studyquery/StudyQueryFactoryTest.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.study;
+package com.yigongil.backend.domain.study.studyquery;
 
 import static org.assertj.core.api.Assertions.assertThat;
